requirements.txt
setuptools==59.5.0
editdistance==0.6.2
matplotlib==3.3.3
numba==0.51.2
numpy==1.20.3
opencv-python==4.5.5.62
pillow==9.0.1
polygon3==3.0.9.1
rapidfuzz==2.13.7
scipy==1.5.2
scikit-image==0.15.0
scikit-learn==1.3.2
shapely==2.0.0
timm==0.5.4
tqdm==4.53.0
bezier==2021.2.12
pandas==2.0.3